Trying adapt to multiprocessing_logging.
